chore: make build and deploy conditional + extract configs

+++ scripts/deploy.sh
@@ -1,17 +1,29 @@
 #!/bin/bash
 set -ev
 
-# Install google cloud SDK and kubectl
-if [[ ! -d "$HOME/google-cloud-sdk/bin" ]]; then rm -rf "$HOME/google-cloud-sdk"; curl https://sdk.cloud.google.com | bash > /dev/null; fi
-source $HOME/google-cloud-sdk/path.bash.inc
-gcloud components update kubectl
-gcloud version
+if [[ ${TEST_BRANCH_NAME} =~ ^release.*$ || ${TEST_BRANCH_NAME} == master ]]
+then
+    # Load configuration
+    . scripts/config.common
+    if [[ ${TEST_BRANCH_NAME} == master ]]
+        then
+            . scripts/config.production
+        else
+            . scripts/config.staging
+    fi
 
-# Authentication process
-gcloud config set account travis@tipi-staging-micro-services.iam.gserviceaccount.com
-gcloud auth activate-service-account --key-file=tipi-staging-micro-services-91d8a0a468e1.json
-gcloud container clusters get-credentials cluster-stg-sydney --zone australia-southeast1-a --project tipi-staging-micro-services
+    # Install google cloud SDK and kubectl
+    if [[ ! -d "$HOME/google-cloud-sdk/bin" ]]; then rm -rf "$HOME/google-cloud-sdk"; curl https://sdk.cloud.google.com | bash > /dev/null; fi
+    source $HOME/google-cloud-sdk/path.bash.inc
+    gcloud components update kubectl
+    gcloud version
 
-# Deploy built version
-SERVICE_VERSION=$(node -p "require('./package.json').version")
-kubectl --record deployment.apps/robots set image deployment.v1.apps/robots robots=rgmax/robots:${SERVICE_VERSION}
+    # Authentication process
+    gcloud config set account ${CLOUD_SERVICE_ACCOUNT_EMAIL}
+    gcloud auth activate-service-account --key-file=${CLOUD_SERVICE_ACCOUNT_KEY_FILE}
+    gcloud container clusters get-credentials ${CLOUD_CLUSTER_NAME} --zone ${CLOUD_CLUSTER_ZONE} --project ${CLOUD_CLUSTER_PROJECT}
+
+    # Deploy built version
+    SERVICE_VERSION=$(node -p "require('./package.json').version")
+    kubectl --record deployment.apps/${APP_DEPLOYMENT_NAME} set image deployment.v1.apps/${APP_DEPLOYMENT_NAME} ${APP_CONTAINER_NAME}=${APP_DOCKER_IMAGE_NAME}:${SERVICE_VERSION}
+fi
